requirements.txt
Flask
request
jsonify
keras
tensorflow
pandas
numpy
newspaper3k
